+++ tests/framework/cli/conftest.py
@@ -23,6 +23,7 @@ from kedro.framework.cli.pipeline import pipeline_cli
 from kedro.framework.cli.project import project_group
 from kedro.framework.cli.registry import registry_cli
 from kedro.framework.cli.starters import create_cli
+from kedro.framework.cli.micropkg import micropkg_cli
 from kedro.framework.project import configure_project, pipelines, settings
 from kedro.framework.startup import ProjectMetadata
 
@@ -105,6 +106,7 @@ def fake_kedro_cli():
             pipeline_cli,
             project_group,
             registry_cli,
+            micropkg_cli,
         ],
     )
